requirements.txt
pydantic>=2.7
PyYAML>=6.0
schedule>=1.2
mysql-connector-python>=9.0
paramiko>=3.4
fastapi>=0.111
uvicorn[standard]>=0.30
openai>=1.34
langgraph>=0.2
tenacity>=8.2
python-dotenv>=1.0
paramiko